Fixed Migration Mismatch, SignUpTopic Spec
- Introduced a model spec file for the SignUpTopic class (currently untested, as our environment has trouble with the permissions for rspec)
- Fixed a type mismatch in our migrations; topic_id in assignment_questionnaires should now be correctly assigned "bigint" everywhere

+++ db/migrate/20251029210525_add_topic_id_to_assignment_questionnaires.rb
@@ -1,6 +1,6 @@
 class AddTopicIdToAssignmentQuestionnaires < ActiveRecord::Migration[8.0]
   def change
-    add_column :assignment_questionnaires, :topic_id, :integer
+    add_column :assignment_questionnaires, :topic_id, :bigint
     add_index :assignment_questionnaires, :topic_id
     add_foreign_key :assignment_questionnaires, :sign_up_topics, column: :topic_id
   end
